+++ luxon.js
@@ -1,4 +1,4 @@
-/* eslint-disable max-classes-per-file */
+/* eslint-disable */
 // these aren't really private, but nor are they really useful to document
 
 /**
