+++ modules/book.js
@@ -1,6 +1,7 @@
-export default class Book {
+class Book {
   constructor(title, author) {
     this.title = title;
     this.author = author;
   }
 }
+export { Book };
