+++ modules/collect.js
@@ -2,7 +2,6 @@ import { Book } from './book.js';
 
 const inputTitle = document.getElementById('title');
 const inputAuthor = document.getElementById('author');
-const submitBtn = document.querySelector('.add-btn');
 const bookSection = document.querySelector('.books');
 class Collect {
   constructor(books = []) {
@@ -52,14 +51,4 @@ class Collect {
     );
   };
 }
-const coll = new Collect();
-if (localStorage.getItem('bookCollection')) {
-  const localBooks = JSON.parse(localStorage.getItem('bookCollection'));
-  localBooks.bookColl.forEach((element) => {
-    coll.add(new Book(element.title, element.author));
-  });
-}
-submitBtn.addEventListener('click', () => {
-  coll.add(new Book(inputTitle.value, inputAuthor.value));
-});
 export { Collect };
